+++ src/opentelemetry-sdk-fastly/FastlySDK.ts
@@ -7,13 +7,12 @@ import { ContextManager, TextMapPropagator, metrics } from "@opentelemetry/api";
 
 import { InstrumentationOption, registerInstrumentations } from "@opentelemetry/instrumentation";
 import { Resource } from "@opentelemetry/resources";
-import { SimpleSpanProcessor, SpanProcessor } from "@opentelemetry/sdk-trace-base";
+import { BatchSpanProcessor, SpanProcessor } from "@opentelemetry/sdk-trace-base";
 import { MeterProvider, MetricReader } from "@opentelemetry/sdk-metrics";
 
 import { FastlySDKConfiguration } from "./types";
-import { FastlySpanProcessor, FastlyTracerConfig, FastlyTracerProvider } from "../opentelemetry-sdk-trace-fastly";
+import { FastlyTracerConfig, FastlyTracerProvider } from "../opentelemetry-sdk-trace-fastly";
 import { setPatchTarget } from "./util";
-import { OTLPExporterFastlyBackendBase } from "../otlp-exporter-fastly-base";
 
 type TransformConfigurationFunction = (configuration: Partial<FastlySDKConfiguration>, event: FetchEvent) => Partial<FastlySDKConfiguration>;
 
@@ -71,13 +70,9 @@ export class FastlySDK {
         tracerProviderConfig.spanLimits = configuration.spanLimits;
       }
 
-      const SpanProcessorClass =
-        configuration.traceExporter instanceof OTLPExporterFastlyBackendBase ?
-          FastlySpanProcessor : SimpleSpanProcessor;
-
       const spanProcessor =
         configuration.spanProcessor ??
-        new SpanProcessorClass(configuration.traceExporter!);
+        new BatchSpanProcessor(configuration.traceExporter!);
 
       this.configureTracerProvider(
         tracerProviderConfig,